+++ tests/browser-automated-tests-playwright/playwright.config.ts
@@ -5,7 +5,6 @@ import { PlaywrightTestConfig } from '@playwright/test';
 const PLAYWRIGHT_HEADLESS = process.env.PLAYWRIGHT_HEADLESS.toLocaleLowerCase() == "true" || false;
 const PLAYWRIGHT_HOST = process.env.PLAYWRIGHT_HOST;
 const PLAYWRIGHT_SLOWMO = parseInt(process.env.PLAYWRIGHT_SLOWMO);
-
 const { devices } = require('@playwright/test');
 /** @type {import('@playwright/test').PlaywrightTestConfig} */
 const config: PlaywrightTestConfig = {
@@ -14,7 +13,6 @@ const config: PlaywrightTestConfig = {
   timeout: 180000,
   retries: 2,
   workers: 3,
-  globalTeardown: './delete-connect-account-id', 
   use: {
 
     baseURL: PLAYWRIGHT_HOST,